rename gallary
