+++ src/modules/products/components/image-gallery/index.tsx
