+++ src/modules/products/templates/index.tsx
@@ -5,7 +5,7 @@ import ProductTabs from "@modules/products/components/product-tabs"
 import RelatedProducts from "@modules/products/components/related-products"
 import ProductInfo from "@modules/products/templates/product-info"
 import React, { useRef } from "react"
-import ImageGallery from "../components/image-gallary"
+import ImageGallery from "../components/image-gallery"
 import MobileActions from "../components/mobile-actions"
 import { PricedProduct } from "@medusajs/medusa/dist/types/pricing"
